+++ tools/testing/test_selections.py
@@ -4,6 +4,7 @@ import os
 import subprocess
 
 from tools.stats.s3_stat_parser import (
+    _get_stripped_CI_job,
     get_previous_reports_for_branch,
     get_previous_reports_for_pr,
     Report, Version2Report,
@@ -20,19 +21,6 @@ class JobTimeJSON(TypedDict):
     commit: str
     JOB_BASE_NAME: str
     job_times: Dict[str, float]
-
-
-def _get_stripped_CI_job() -> str:
-    """E.g. convert 'pytorch_windows_vs2019_py36_cuda10.1_build' to 'pytorch_windows_vs2019_py36_cuda10.1'.
-    """
-    job = os.environ.get("JOB_BASE_NAME", "").rstrip('0123456789')
-    if job.endswith('_slow_test'):
-        job = job[:len(job) - len('_slow_test')]
-    elif job.endswith('_test') or job.endswith('-test'):
-        job = job[:len(job) - len('_test')]
-    elif job.endswith('_build') or job.endswith('-build'):
-        job = job[:len(job) - len('_build')]
-    return job
 
 
 def _get_job_times_json(job_times: Dict[str, float]) -> JobTimeJSON:
